initial integration of the s3 transfer manager
This replaces the naive multipart upload implementation with the new
S3 transfer manager. Unfortunately, in order for the transfer manager
to do multipart uploads, it must use the CRT client, but the CRT
client does not currently allow you to specify path-style access,
which is a requirement for some non-AWS S3 implementations. Hopefully
they add support soon.